Added a Landing page
Landing page that describes what the site is. Also redirects a logged-out user to the landing page instead of the dashboard.  Updated the login to be /login so / is the landing page. Added a basic logout button. Also updated the dashboard to fix the card layout issues. Navbar updates based on login status.